requirements.txt
Desktopmagic==14.3.11
pywin32==227
Pillow==6.2.1
google-api-core==1.14.3